requirements.txt
attrs==25.3.0
blinker==1.9.0
brother_ql-inventree==1.3
click==8.2.1
dominate==2.9.1
Flask==3.1.2
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
packbits==0.6
pdf2image==1.17.0
pillow==11.3.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.2
python-barcode==0.16.1
pyusb==1.3.1
fonttools==4.59.2
qrcode==8.2
visitor==0.1.3
Werkzeug==3.1.3
